Separate KMS client for datasets
